# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmergencyInput {
  /** The user's spoken or typed description of the emergency */
  transcript: string;
  /**
     * ISO language code (e.g. "en", "hi", "ta")
     * @nullable
     */
  language?: string | null;
}

/**
 * Urgency level
 */
export type EmergencyAnalysisUrgency = typeof EmergencyAnalysisUrgency[keyof typeof EmergencyAnalysisUrgency];


export const EmergencyAnalysisUrgency = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

/**
 * Inferred context about the person in distress based on relational cues in the transcript
 * @nullable
 */
export type EmergencyAnalysisSubjectContext = {
  /** The relationship word detected (e.g. "father", "wife", "son") */
  relationship: string;
  estimatedGender: 'male' | 'female' | 'unknown';
  /** Human-readable estimated age range (e.g. "55–70 years") */
  estimatedAgeRange: string;
  confidence: 'high' | 'medium' | 'low';
} | null;

export type ActionItemType = typeof ActionItemType[keyof typeof ActionItemType];


export const ActionItemType = {
  call: 'call',
  locate: 'locate',
  guide: 'guide',
  contact: 'contact',
} as const;

export interface ActionItem {
  id: string;
  label: string;
  /** Icon name (e.g. "phone", "map-pin", "mic") */
  icon: string;
  type: ActionItemType;
  /** @nullable */
  phoneNumber?: string | null;
}

export interface GuidanceStep {
  order: number;
  instruction: string;
  icon: string;
  /** @nullable */
  completed?: boolean | null;
}

export interface Helpline {
  id: string;
  name: string;
  number: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  available247?: boolean | null;
}

export interface EmergencyAnalysis {
  /** Detected emergency category (e.g. "medical", "financial", "crime", "disaster") */
  category: string;
  /** Urgency level */
  urgency: EmergencyAnalysisUrgency;
  /** Short label for the detected emergency (e.g. "Medical Emergency Detected") */
  title: string;
  /** Brief reassuring explanation of what JivAI understood */
  summary: string;
  /**
     * Inferred context about the person in distress based on relational cues in the transcript
     * @nullable
     */
  subjectContext?: EmergencyAnalysisSubjectContext;
  actions: ActionItem[];
  /** @nullable */
  guidanceSteps?: GuidanceStep[] | null;
  /** @nullable */
  helplines?: Helpline[] | null;
}

export interface GuidanceResponse {
  category: string;
  title: string;
  steps: GuidanceStep[];
}

export interface EmergencyCategory {
  id: string;
  label: string;
  icon: string;
  description: string;
  /** @nullable */
  color?: string | null;
}

export type ListHelplinesParams = {
category?: string;
};