# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JivAI Emergency Crisis Response API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: emergency
    description: Emergency session and classification
  - name: guidance
    description: Guided response steps and helplines
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /emergency/analyze:
    post:
      operationId: analyzeEmergency
      tags: [emergency]
      summary: Analyze emergency from text or transcript
      description: Classifies emergency type, urgency, and returns immediate action options
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyInput"
      responses:
        "200":
          description: Emergency analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAnalysis"

  /emergency/guidance/{category}:
    get:
      operationId: getGuidance
      tags: [guidance]
      summary: Get step-by-step guidance for an emergency category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guided response steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuidanceResponse"

  /emergency/helplines:
    get:
      operationId: listHelplines
      tags: [guidance]
      summary: List emergency helplines
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of emergency helplines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Helpline"

  /emergency/categories:
    get:
      operationId: listCategories
      tags: [emergency]
      summary: List all supported emergency categories
      responses:
        "200":
          description: Supported emergency categories with descriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyCategory"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    EmergencyInput:
      type: object
      required:
        - transcript
      properties:
        transcript:
          type: string
          description: The user's spoken or typed description of the emergency
        language:
          type: ["string", "null"]
          description: ISO language code (e.g. "en", "hi", "ta")

    EmergencyAnalysis:
      type: object
      required:
        - category
        - urgency
        - title
        - actions
        - summary
      properties:
        category:
          type: string
          description: Detected emergency category (e.g. "medical", "financial", "crime", "disaster")
        urgency:
          type: string
          enum: [critical, high, medium, low]
          description: Urgency level
        title:
          type: string
          description: Short label for the detected emergency (e.g. "Medical Emergency Detected")
        summary:
          type: string
          description: Brief reassuring explanation of what JivAI understood
        subjectContext:
          type: ["object", "null"]
          description: Inferred context about the person in distress based on relational cues in the transcript
          properties:
            relationship:
              type: string
              description: The relationship word detected (e.g. "father", "wife", "son")
            estimatedGender:
              type: string
              enum: [male, female, unknown]
            estimatedAgeRange:
              type: string
              description: Human-readable estimated age range (e.g. "55–70 years")
            confidence:
              type: string
              enum: [high, medium, low]
          required:
            - relationship
            - estimatedGender
            - estimatedAgeRange
            - confidence
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionItem"
        guidanceSteps:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/GuidanceStep"
        helplines:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/Helpline"

    ActionItem:
      type: object
      required:
        - id
        - label
        - icon
        - type
      properties:
        id:
          type: string
        label:
          type: string
        icon:
          type: string
          description: Icon name (e.g. "phone", "map-pin", "mic")
        type:
          type: string
          enum: [call, locate, guide, contact]
        phoneNumber:
          type: ["string", "null"]

    GuidanceResponse:
      type: object
      required:
        - category
        - title
        - steps
      properties:
        category:
          type: string
        title:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/GuidanceStep"

    GuidanceStep:
      type: object
      required:
        - order
        - instruction
        - icon
      properties:
        order:
          type: integer
        instruction:
          type: string
        icon:
          type: string
        completed:
          type: ["boolean", "null"]

    Helpline:
      type: object
      required:
        - id
        - name
        - number
        - category
      properties:
        id:
          type: string
        name:
          type: string
        number:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        available247:
          type: ["boolean", "null"]

    EmergencyCategory:
      type: object
      required:
        - id
        - label
        - icon
        - description
      properties:
        id:
          type: string
        label:
          type: string
        icon:
          type: string
        description:
          type: string
        color:
          type: ["string", "null"]