# lib/api-zod/src/generated/types/actionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItemType } from './actionItemType';

export interface ActionItem {
  id: string;
  label: string;
  /** Icon name (e.g. "phone", "map-pin", "mic") */
  icon: string;
  type: ActionItemType;
  /** @nullable */
  phoneNumber?: string | null;
}