# lib/api-zod/src/generated/types/emergencyAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItem } from './actionItem';
import type { EmergencyAnalysisSubjectContext } from './emergencyAnalysisSubjectContext';
import type { EmergencyAnalysisUrgency } from './emergencyAnalysisUrgency';
import type { GuidanceStep } from './guidanceStep';
import type { Helpline } from './helpline';

export interface EmergencyAnalysis {
  /** Detected emergency category (e.g. "medical", "financial", "crime", "disaster") */
  category: string;
  /** Urgency level */
  urgency: EmergencyAnalysisUrgency;
  /** Short label for the detected emergency (e.g. "Medical Emergency Detected") */
  title: string;
  /** Brief reassuring explanation of what JivAI understood */
  summary: string;
  /**
     * Inferred context about the person in distress based on relational cues in the transcript
     * @nullable
     */
  subjectContext?: EmergencyAnalysisSubjectContext;
  actions: ActionItem[];
  /** @nullable */
  guidanceSteps?: GuidanceStep[] | null;
  /** @nullable */
  helplines?: Helpline[] | null;
}