# lib/api-zod/src/generated/types/emergencyCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyCategory {
  id: string;
  label: string;
  icon: string;
  description: string;
  /** @nullable */
  color?: string | null;
}