# lib/api-zod/src/generated/types/guidanceResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */
import type { GuidanceStep } from './guidanceStep';

export interface GuidanceResponse {
  category: string;
  title: string;
  steps: GuidanceStep[];
}