# lib/api-zod/src/generated/types/helpline.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JivAI Emergency Crisis Response API
 * OpenAPI spec version: 0.1.0
 */

export interface Helpline {
  id: string;
  name: string;
  number: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  available247?: boolean | null;
}